types and clone
